Added altair/util/log.py as a logging class. Modified JSONExtractor to incorporate logging messages.

+++ altair/preprocess00/JSONExtractor.py
@@ -1,6 +1,11 @@
 import gzip
 import json
 import os
+import time
+
+from altair.util.log import getLogger
+
+logger = getLogger(__name__)
 
 class JSONExtractor:
     def __init__(self, output_dir, output_extension):
@@ -17,13 +22,17 @@ class JSONExtractor:
             for line in f1:
                 j = json.loads(line.decode("utf-8"))
                 target = os.path.join(folder, "%s.%s" % (j["id"], self.output_extension))
-                with open(target, "wb") as f2:
-                    f2.write(j["content"].encode("utf-8"))
+                if not os.path.exists(target):
+                    with open(target, "wb") as f2:
+                        f2.write(j["content"].encode("utf-8"))
 
     def extract_dir(self, input_dir):
         for json_gz_file in sorted(os.listdir(input_dir)):
             fullpath = os.path.join(input_dir, json_gz_file)
+            logger.info("Processing file: %s" % fullpath)
+            start_time = time.time()
             self.extract(fullpath)
+            logger.info("Elapsed time: %s" % (time.time()-start_time))
 
 if __name__ == "__main__":
     import argparse
